fix: try to fix issue templates not applied issue
Previously, the issue templates are configured with yml,
but these templates are not applied when creating new a issue.
Since I cannot find a way to debug the template issue,
So this commit changed to use markdown format,
which is more intuitive to write and are also used by many other popular open source projects.